fix Bolt11Invoice.toString invalid call that was causing many tests to fail.
now tests should be good again.

except for the fact that some trampoline/private-routing-related tests are failing sometimes, sometimes not.
I don't understand that part of the code enough to debug for now, and it seems to be some sort of race condition,
I don't know. it's not very important so I'll postpone it for later.